clients/upssched.c: fix receiver of strtol()

+++ clients/upssched.c
@@ -180,7 +180,7 @@ static void checktimers(void)
 static void start_timer(const char *name, const char *ofsstr)
 {
 	time_t	now;
-	int	ofs;
+	long	ofs;
 	ttype_t	*tmp, *last;
 
 	/* get the time */
@@ -195,7 +195,7 @@ static void start_timer(const char *name, const char *ofsstr)
 	}
 
 	if (verbose)
-		upslogx(LOG_INFO, "New timer: %s (%d seconds)", name, ofs);
+		upslogx(LOG_INFO, "New timer: %s (%ld seconds)", name, ofs);
 
 	/* now add to the queue */
 	tmp = last = thead;
